refactor(webdav): merge error_handler.go into file_system.go and simplify
- Move HTTPError, customErrorHandler, and errorHandlingFile into file_system.go
- Eliminate 4 passthrough methods by embedding FileSystem in customErrorHandler
- Remove dead deadProps code path in propfind/prop.go
- Simplify handlePropfindError and context chain in adapter.go
