+++ internal/webdav/monitored_file_system.go
@@ -84,7 +84,3 @@ func (m *monitoredFile) Write(p []byte) (n int, err error) {
 	}
 	return m.File.Write(p)
 }
-
-func (m *monitoredFile) Close() error {
-	return m.File.Close()
-}
